+++ MoreShipUpgrades/API/HunterSamples.cs
@@ -1,7 +1,6 @@
 ﻿using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.Misc;
 using MoreShipUpgrades.UpgradeComponents.Items;
-using System;
 using System.Collections.Generic;
 using UnityEngine;